DP-3911 - Form Inputs JS - renamed from textarea.js
